+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased][unreleased]
 
 - Fixed API endpoints local queue settings applying
+- Reimplement global timeouts.request usage during a Procedure invocation
 
 ## [3.0.13][] - 2023-10-22